Moved custom implementations to their own folder, separated the user service and user details service classes
