+++ src/main/java/com/magvy/experis/javalava_backend/controllers/CommentController.java
@@ -2,7 +2,7 @@ package com.magvy.experis.javalava_backend.controllers;
 
 import com.magvy.experis.javalava_backend.application.DTOs.incoming.CommentDTORequest;
 import com.magvy.experis.javalava_backend.application.DTOs.outgoing.CommentDTOResponse;
-import com.magvy.experis.javalava_backend.application.security.config.CustomUserDetails;
+import com.magvy.experis.javalava_backend.application.security.config.custom.CustomUserDetails;
 import com.magvy.experis.javalava_backend.controllers.util.ResponseUtil;
 import com.magvy.experis.javalava_backend.domain.services.CommentService;
 import org.springframework.http.ResponseEntity;
